requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-filter==24.2 ; python_version >= "3.10" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.27.2 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.22.0 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.6 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg==3.1.19 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
